+++ src/types.d.ts
@@ -243,7 +243,7 @@ interface VisitorReducer {
   usedOds: OdPlace[]
 }
 
-interface MenuReducer {
+interface UiReducer {
   info: boolean,
   pathPanel: boolean,
   pathPanelContent: string | null,
@@ -259,5 +259,5 @@ interface ReduxState {
   destination: DestinationReducer,
   mapPopup: MapPopupReducer,
   visitor: VisitorReducer,
-  menu: MenuReducer
+  ui: UiReducer
 }
